luppp: updated page
Refactored the lines to be < 80 characters (easier to read here :)
and fixed a few typos. I know the diff here looks huge, but I didn't
change that much of your content - its just all line wraps.

I added a few screenshots of the routing, and changed the Error: to
Warning: messages, and linked to the issue that is trackcing that bug.